modelo_IA/requirements.txt
flask==2.3.3
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3
scipy==1.10.1
joblib==1.3.2
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
threading
gunicorn==21.2.0
Flask-CORS==4.0.0